Removed global variable NMAX
This required defining a Fullerene class (instead of a struct) with vector
data structures for 'primal', 'dual', and 'edges', changing
'fvars' and 'evars' (in lp.cpp) to vectors, and updating the arguments of
the functions that used these data structures. Additionally, instead
of files being saved as nnn_pp_{rest of filename}, which forces the
maximum number of vertices of a fullerene read-in to be 998 (since it
only provides three digits for the filename), to pp_{rest of filename}.
Files in output/ now reflect this naming convention.
Readme was also updated to reflect these changes.
Fixes #3.